iot/requirements.txt
Django==2.1.15
django-mssql==1.8
djangorestframework==3.14.0
